Switch Time::elapsedMsecs to return qint64.
When dealing with elapsed times you almost always want signed integers
because zero crossings can cause crazy large numbers.

For unsigned integers: (small number - small number) => GIGANTIC NUMBER
For signed integers: (small number - small number) = +/- small number

+++ src/controllers/controllerengine.h
@@ -187,7 +187,7 @@ class ControllerEngine : public QObject {
     // 256 (default) available virtual decks is enough I would think.
     //  If more are needed at run-time, these will move to the heap automatically
     QVarLengthArray<int> m_intervalAccumulator;
-    QVarLengthArray<uint> m_lastMovement;
+    QVarLengthArray<qint64> m_lastMovement;
     QVarLengthArray<double> m_dx, m_rampTo, m_rampFactor;
     QVarLengthArray<bool> m_ramp, m_brakeActive;
     QVarLengthArray<AlphaBetaFilter*> m_scratchFilters;
